+++ model/internal/cmd/pdatagen/internal/trace_structs.go
@@ -194,10 +194,6 @@ var spanStatus = &messageValueStruct{
 			rawType:         "otlptrace.Status_StatusCode",
 			defaultVal:      "StatusCode(0)",
 			testVal:         "StatusCode(1)",
-			// Generate code without setter. Setter will be manually coded since we
-			// need to also change DeprecatedCode when Code is changed according
-			// to OTLP spec https://github.com/open-telemetry/opentelemetry-proto/blob/59c488bfb8fb6d0458ad6425758b70259ff4a2bd/opentelemetry/proto/trace/v1/trace.proto#L231
-			manualSetter: true,
 		},
 		&primitiveField{
 			fieldName:       "Message",