+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 🛑 Breaking changes 🛑
+
+- Updated to OTLP 0.12.0. Deprecated traces and metrics messages that existed
+  in 0.11.0 are no longer converted to the messages and fields that replaced the deprecated ones.
+  Received deprecated messages and fields will be now ignored. In OTLP/JSON in the
+  instrumentationLibraryLogs object the "logs" field is now named "logRecords" (#4724)
+
 ## v0.43.1 Beta
 
 ## 🧰 Bug fixes 🧰